requirements_full.txt
streamlit
pandas
google-generativeai
python-dotenv
gspread
google-auth
